# openapi.json
{
  "components": {
    "schemas": {
...
      "DiffOut": {
        "properties": {
          "model_version_id": {
            "title": "Model Version Id",
            "type": "string"
          },
          "new": {
            "title": "New",
            "type": "integer"
          },
          "persisting": {
            "title": "Persisting",
            "type": "integer"
          },
          "previous_id": {
            "title": "Previous Id",
            "type": "string"
          },
          "proposal_score": {
            "additionalProperties": true,
            "title": "Proposal Score",
            "type": "object"
          },
          "regressed": {
            "title": "Regressed",
            "type": "integer"
          },
          "resolved": {
            "title": "Resolved",
            "type": "integer"
          }
        },
        "required": [
          "model_version_id",
          "previous_id",
          "resolved",
          "regressed",
          "new",
          "persisting",
          "proposal_score"
        ],
        "title": "DiffOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelVersionOut": {
        "properties": {
          "element_count": {
            "title": "Element Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ifc_sha256": {
            "title": "Ifc Sha256",
            "type": "string"
          },
          "ingested_at": {
            "format": "date-time",
            "title": "Ingested At",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "speckle_stream": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speckle Stream"
          },
          "stats": {
            "additionalProperties": true,
            "title": "Stats",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "ifc_sha256",
          "element_count",
          "status",
          "ingested_at"
        ],
        "title": "ModelVersionOut",
        "type": "object"
      },
      "ProjectIn": {
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectIn",
        "type": "object"
      },
      "ProjectOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nJudges MEP coordination on an IFC model, proposes moves that cite a clause, and\nverifies every one of them through three independent monitors before a reviewer\never sees it.\n\nThe model of record is never written to. Proposals are vectors and clause\nreferences, applied by an engineer in the authoring tool; the original file's\nsha256 is reported before and after every run so the promise is evidence rather\nthan a claim.\n",
    "title": "mep-judge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/events/{topic}": {
      "get": {
        "operationId": "event_stream_events__topic__get",
        "parameters": [
          {
            "in": "path",
            "name": "topic",
            "required": true,
            "schema": {
              "title": "Topic",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Stream",
        "tags": [
          "events"
        ]
      }
    },
...
    "/models/{version_id}": {
      "get": {
        "operationId": "get_model_models__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{version_id}/diff/{previous_id}": {
      "get": {
        "operationId": "diff_models_models__version_id__diff__previous_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "previous_id",
            "required": true,
            "schema": {
              "title": "Previous Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Models",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{version_id}/zones": {
      "get": {
        "operationId": "list_zones_models__version_id__zones_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ZoneOut"
                  },
                  "title": "Response List Zones Models  Version Id  Zones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Zones",
        "tags": [
          "zones"
        ]
      }
    },
    "/projects": {
      "post": {
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/keys": {
      "post": {
        "operationId": "create_key_projects__project_id__keys_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_key_projects__project_id__keys_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyIssued"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Key",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/models": {
      "post": {
        "operationId": "upload_model_projects__project_id__models_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_model_projects__project_id__models_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Model",
        "tags": [
          "models"
        ]
      }
    },
    "/zones/{zone_id}/bcf.zip": {
      "get": {
        "operationId": "get_bcf_zones__zone_id__bcf_zip_get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bcf",
        "tags": [
          "review"
        ]
      }
    },
    "/zones/{zone_id}/change_set.json": {
      "get": {
        "operationId": "get_change_set_zones__zone_id__change_set_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Set",
        "tags": [
          "review"
        ]
      }
    },
    "/zones/{zone_id}/clashes": {
      "get": {
        "operationId": "list_clashes_zones__zone_id__clashes_get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClashOut"
                  },
                  "title": "Response List Clashes Zones  Zone Id  Clashes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clashes",
        "tags": [
          "zones"
        ]
      }
    },
    "/zones/{zone_id}/proposals": {
      "get": {
        "operationId": "list_proposals_zones__zone_id__proposals_get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalOut"
                  },
                  "title": "Response List Proposals Zones  Zone Id  Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals",
        "tags": [
          "zones"
        ]
      }
    },
    "/zones/{zone_id}/review": {
      "post": {
        "operationId": "review_zone_zones__zone_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Zone",
        "tags": [
          "review"
        ]
      }
    }
  },
...
}